try again

+++ webview/build_webview_with_qt5.sh
@@ -14,8 +14,8 @@ QT_VERSION="$QT_MAJOR.$QT_MINOR.$QT_BUG_FIX"
 DEBIAN_VERSION=$(lsb_release -cs)
 MAKE_CORES="$(expr $(nproc) + 2)"
 
-SCREENCONNECT_RELEASE_URL="https://github.com/Copper-Clock/Screenconnect/releases"
-WEBVIEW_VERSION="0.3.3"
+ANTHIAS_RELEASE_URL="https://github.com/Copper-Clock/Screenconnect/releases"
+WEBVIEW_VERSION="0.3.5"
 
 mkdir -p "$BUILD_TARGET"
 mkdir -p "$SRC"
@@ -38,11 +38,19 @@ function download_and_extract_qt5() {
     local SRC_DIR="$1"
     local DEVICE="$2"
 
-    WEBVIEW_DL_URL="$SCREENCONNECT_RELEASE_URL/download/WebView-v$WEBVIEW_VERSION/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz"
+    WEBVIEW_DL_URL="$ANTHIAS_RELEASE_URL/download/WebView-v$WEBVIEW_VERSION/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz"
     WEBVIEW_DL_URL_SHA256="$WEBVIEW_DL_URL.sha256"
 
-    curl -sL "$WEBVIEW_DL_URL" -o /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz
-    curl -sL "$WEBVIEW_DL_URL_SHA256" -o /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz.sha256
+    if [ ! -f /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz ]; then
+        curl -sL "$WEBVIEW_DL_URL" -o /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz
+    fi
+
+    if [ ! -f /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz.sha256 ]; then
+        curl -sL "$WEBVIEW_DL_URL_SHA256" -o /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz.sha256
+    fi
+
+    cp -n /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz /build/
+    cp -n /tmp/qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz.sha256 /build/
 
     cd /tmp
     sha256sum -c "qt5-$QT_VERSION-$DEBIAN_VERSION-$DEVICE.tar.gz.sha256"
@@ -65,9 +73,9 @@ function build_qt () {
     make -j"$MAKE_CORES"
     make install
 
-    mkdir -p fakeroot/bin fakeroot/share/TccconnectWebview
-    mv TccconnectWebview fakeroot/bin/
-    cp -rf /webview/res fakeroot/share/TccconnectWebview/
+    mkdir -p fakeroot/bin fakeroot/share/ScreenlyWebview
+    mv ScreenlyWebview fakeroot/bin/
+    cp -rf /webview/res fakeroot/share/ScreenlyWebview/
 
     pushd fakeroot
     tar cfz "$BUILD_TARGET/webview-$QT_VERSION-$DEBIAN_VERSION-$1-$GIT_HASH.tar.gz" .
@@ -82,7 +90,7 @@ fetch_cross_compile_tool
 
 if [ ! "${TARGET-}" ]; then
     # Let's work our way through all Pis in order of relevance
-    for device in pi4; do
+    for device in pi4 pi3 pi2 pi1; do
         build_qt "$device"
     done
 else
